libretro: fix linker error with tests
